+++ sources/Device/src/Settings/SettingsTime.cpp
@@ -33,7 +33,7 @@ int sTime_TPosInBytes()
         {{0, 8192, 16382}, {0, 8192, 16382}},
         {{0, 16384, 32766}, {0, 16384, 32766}}
     };
-    return m[FPGA_ENUM_POINTS][SET_PEAKDET][TPOS];
+    return m[(int)FPGA_ENUM_POINTS][SET_PEAKDET][TPOS];
 }
 
 
@@ -50,7 +50,7 @@ int sTime_TPosInPoints()
         {{0, 8192, 16382}, {0, 4096, 8190}},
         {{0, 16384, 32766}, {0, 8192, 16382}}
     };
-    return m[FPGA_ENUM_POINTS][SET_PEAKDET][TPOS];
+    return m[(int)FPGA_ENUM_POINTS][SET_PEAKDET][TPOS];
 }
